feat: chrome extension build
Closes: #1383
