+++ packages/js/package.json
@@ -69,6 +69,7 @@
     "jest-fetch-mock": "^3.0.3",
     "nock": "^13.2.1",
     "rollup": "^2.77.0",
+    "rollup-plugin-strip-code": "^0.2.7",
     "rollup-plugin-terser": "^7.0.2",
     "sinon": "^14.0.0",
     "supertest": "^6.1.6",
