+++ packages/js/rollup.config.js
@@ -2,6 +2,7 @@ import replace from '@rollup/plugin-replace'
 import resolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
 import { terser } from 'rollup-plugin-terser'
+import stripCode from 'rollup-plugin-strip-code'
 import pkg from './package.json'
 
 const sharedPlugins = [
@@ -41,6 +42,25 @@ export default [
       sourcemap: true
     },
     plugins: [...sharedPlugins, terser()]
+  },
+
+  // Chrome Extension build (minified) - https://github.com/honeybadger-io/honeybadger-js/issues/1383
+  {
+    input: 'build/src/browser.js',
+    output: {
+      name: 'Honeybadger',
+      file: 'dist/browser/honeybadger.ext.min.js',
+      format: 'umd',
+      sourcemap: true
+    },
+    plugins: [
+      stripCode({
+        start_comment: 'ROLLUP_STRIP_CODE_CHROME_EXTENSION_START',
+        end_comment: 'ROLLUP_STRIP_CODE_CHROME_EXTENSION_END'
+      }),
+      ...sharedPlugins,
+      terser()
+    ]
   },
 
   // Server build
